Move `ConstOrExpr` into common code
